requirements.txt
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
# Real-Time Stock Portfolio Optimizer
# Production Requirements — NSE/BSE Edition
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━

# Core Data Science
yfinance>=0.2.28
pandas>=2.0.3
numpy>=1.24.4
scipy>=1.11.4
scikit-learn>=1.3.2
pyarrow>=13.0.0
statsmodels>=0.14.0

# Deep Learning (LSTM)
tensorflow>=2.13.0
keras>=2.13.1

# Time Series Forecasting
prophet>=1.1.5

# Reinforcement Learning
stable-baselines3>=2.2.1
gymnasium>=0.29.1
torch>=2.1.0

# MLOps & Experiment Tracking
mlflow>=2.9.2
evidently>=0.4.22

# API Layer
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
redis>=5.0.1
slowapi>=0.1.9
httpx>=0.25.2
python-multipart>=0.0.6
aiofiles>=23.2.1

# Dashboard
streamlit>=1.28.2
plotly>=5.18.0
matplotlib>=3.8.2
seaborn>=0.13.0

# Feature Importance
shap>=0.43.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.2
schedule>=1.2.1
requests>=2.31.0
aiohttp>=3.9.1
tqdm>=4.66.1
joblib>=1.3.2

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0

# Code Quality
flake8>=6.1.0
black>=23.11.0
isort>=5.12.0

# Data Validation
pydantic-settings>=2.1.0
